Converted the x an y coordinate parameters of ReadElem and slotReadElem to doubles giving better precision when copying and pasting objects.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@3743 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribus.h
@@ -464,7 +464,7 @@ public slots:
 	void setItemHoch(int h);
 	void setStilvalue(int s);
 	void setItemTypeStyle(int id);
-	void slotElemRead(QString Name, int x, int y, bool art, bool loca, ScribusDoc* docc, ScribusView* vie);
+	void slotElemRead(QString Name, double x, double y, bool art, bool loca, ScribusDoc* docc, ScribusView* vie);
 	void slotChangeUnit(int art, bool draw = true);
 	void NoFrameEdit();
 	void ApplyMasterPage();
